+++ packages/webapp/src/store/hooks.ts
@@ -1,8 +1,0 @@
-import { createTypedHooks } from 'easy-peasy';
-import { StoreModel } from './store';
-
-const typedHooks = createTypedHooks<StoreModel>();
-
-export const useStoreActions = typedHooks.useStoreActions;
-export const useStoreDispatch = typedHooks.useStoreDispatch;
-export const useStoreState = typedHooks.useStoreState;